LINK: Should fix build with non-GCC compilers
git-svn-id: https://svn.code.sf.net/p/vbam/code/branches/bgk-link@1142 a31d4220-a93d-0410-bf67-fe4944624d44

+++ src/gba/GBALink.cpp
@@ -1297,7 +1297,7 @@ ConnectionState InitLink(LinkMode mode)
 
 	// Find the link driver
 	linkDriver = NULL;
-	for (uint i = 0; i < sizeof(linkDrivers) / sizeof(linkDrivers[0]); i++) {
+	for (u8 i = 0; i < sizeof(linkDrivers) / sizeof(linkDrivers[0]); i++) {
 		if (linkDrivers[i].mode == mode) {
 			linkDriver = &linkDrivers[i];
 			break;
